+++ tests/analysis/polarization/test_polarization.py
@@ -161,15 +161,15 @@ def test_polarization_with_differentially_polarized_markers():
             0: {
                 "marker": "A",
                 "morans_i": 0.3009109262187527,
-                "morans_z": 3.6409882113983114,
-                "morans_p_value": 0.00013579678673041425,
+                "morans_z": 8.855302575253617,
+                "morans_p_value": 4.1728555387292095e-19,
                 "component": "PXLCMP0000000",
             },
             1: {
                 "marker": "C",
                 "morans_i": -0.001864280387770322,
-                "morans_z": -2.558088675046134,
-                "morans_p_value": 0.005262462500942811,
+                "morans_z": 0.33792334994637657,
+                "morans_p_value": 0.3677104754311888,
                 "component": "PXLCMP0000000",
             },
         },
@@ -223,15 +223,15 @@ def test_polarization_with_min_marker_count():
             0: {
                 "marker": "A",
                 "morans_i": 0.3009109262187527,
-                "morans_z": 3.6409882113983114,
-                "morans_p_value": 0.00013579678673041425,
+                "morans_z": 6.6115290465881955,
+                "morans_p_value": 1.9018523168814586e-11,
                 "component": "PXLCMP0000000",
             },
             1: {
                 "marker": "C",
                 "morans_i": -0.001864280387770322,
-                "morans_z": -2.558088675046134,
-                "morans_p_value": 0.005262462500942811,
+                "morans_z": -0.11473061339664403,
+                "morans_p_value": 0.4543293240814379,
                 "component": "PXLCMP0000000",
             },
         },
@@ -265,9 +265,169 @@ def test_permuted_polarization_with_differentially_polarized_markers():
     random_vertex["markers"]["C"] = 10
 
     scores = polarization_scores_component_graph(
-        graph, component_id="PXLCMP0000000", n_permutations=10, random_seed=1
+        graph,
+        component_id="PXLCMP0000000",
+        n_permutations=10,
+        random_seed=1,
+        min_marker_count=0,
     )
 
+    # We don't expect to get a value for B, since it has only one value in it.
+    # Hence it is filtered out.
+    expected = pd.DataFrame.from_dict(
+        {
+            0: {
+                "marker": "A",
+                "morans_i": 0.3009109262187527,
+                "morans_z": 8.855302575253617,
+                "morans_p_value": 4.1728555387292095e-19,
+                "component": "PXLCMP0000000",
+            },
+            1: {
+                "marker": "C",
+                "morans_i": -0.001864280387770322,
+                "morans_z": 0.33792334994637657,
+                "morans_p_value": 0.3677104754311888,
+                "component": "PXLCMP0000000",
+            },
+        },
+        orient="index",
+    )
+    # test polarization scores
+    assert_frame_equal(scores, expected, check_exact=False, atol=1e-3)
+
+
+def test_polarization_transformation():
+    # Set seed to get same graph every time
+    graph = create_randomly_connected_bipartite_graph(
+        n1=50, n2=100, p=0.1, random_seed=2
+    )
+
+    rng = np.random.default_rng(1)
+
+    for v in graph.vs:
+        v["markers"] = {"A": 0, "B": 1, "C": 0, "D": 0}
+    random_vertex = graph.vs.get_vertex(
+        rng.integers(low=0, high=graph.vcount(), size=1)[0]
+    )
+    random_vertex["markers"]["A"] = 5
+    neighbors = random_vertex.neighbors()
+    for n in neighbors:
+        n["markers"]["A"] = 2
+
+    random_vertex = graph.vs.get_vertex(
+        rng.integers(low=0, high=graph.vcount(), size=1)[0]
+    )
+    random_vertex["markers"]["C"] = 10
+
+    random_vertex = graph.vs.get_vertex(
+        rng.integers(low=0, high=graph.vcount(), size=1)[0]
+    )
+    random_vertex["markers"]["D"] = 4
+
+    scores_1 = polarization_scores_component_graph(
+        graph,
+        component_id="PXLCMP0000000",
+        n_permutations=10,
+        random_seed=1,
+        min_marker_count=5,
+    )
+
+    scores_2 = polarization_scores_component_graph(
+        graph,
+        component_id="PXLCMP0000000",
+        n_permutations=10,
+        random_seed=1,
+        min_marker_count=0,
+    )
+
+    # We don't expect to get a value for B, since it has only one value in it.
+    # We don't expect to get a value for D, since the marker is filtered due to
+    # low counts
+    expected_1 = pd.DataFrame.from_dict(
+        {
+            0: {
+                "marker": "A",
+                "morans_i": 0.3009109262187527,
+                "morans_z": 6.6115290465881955,
+                "morans_p_value": 1.9018523168814586e-11,
+                "component": "PXLCMP0000000",
+            },
+            1: {
+                "marker": "C",
+                "morans_i": -0.001864280387770322,
+                "morans_z": -0.11473061339664403,
+                "morans_p_value": 0.4543293240814379,
+                "component": "PXLCMP0000000",
+            },
+        },
+        orient="index",
+    )
+
+    # We don't expect to get a value for B, since it has only one value in it.
+    # We DO expect to get a value for D, since it is not filtered this time
+    expected_2 = pd.DataFrame.from_dict(
+        {
+            0: {
+                "marker": "A",
+                "morans_i": 0.3009109262187527,
+                "morans_z": 6.6115290465881955,
+                "morans_p_value": 1.9018523168814586e-11,
+                "component": "PXLCMP0000000",
+            },
+            1: {
+                "marker": "C",
+                "morans_i": -0.001864280387770322,
+                "morans_z": -0.11473061339664403,
+                "morans_p_value": 0.4543293240814379,
+                "component": "PXLCMP0000000",
+            },
+            2: {
+                "marker": "D",
+                "morans_i": -0.0031455494542742875,
+                "morans_z": 0.8264801996063112,
+                "morans_p_value": 0.204265872789349,
+                "component": "PXLCMP0000000",
+            },
+        },
+        orient="index",
+    )
+
+    # test polarization scores
+    assert_frame_equal(scores_1, expected_1, check_exact=False, atol=1e-3)
+    assert_frame_equal(scores_2, expected_2, check_exact=False, atol=1e-3)
+
+    # we also expect the scores to be the same for A and B
+    assert_frame_equal(scores_1, scores_2.head(2), check_exact=False, atol=1e-3)
+
+
+def test_polarization_backward_compatibility():
+    # This tests that we get the same polarity score as the original
+    # polarization score where markers were filtered prior to CLR
+    # transformation.
+
+    # Set seed to get same graph every time
+    graph = create_randomly_connected_bipartite_graph(
+        n1=50, n2=100, p=0.1, random_seed=2
+    )
+    rng = np.random.default_rng(1)
+
+    for v in graph.vs:
+        v["markers"] = {"A": 0, "B": 0, "C": 0}
+    random_vertex = graph.vs.get_vertex(
+        rng.integers(low=0, high=graph.vcount(), size=1)[0]
+    )
+    random_vertex["markers"]["A"] = 5
+    neighbors = random_vertex.neighbors()
+    for n in neighbors:
+        n["markers"]["A"] = 2
+    random_vertex = graph.vs.get_vertex(
+        rng.integers(low=0, high=graph.vcount(), size=1)[0]
+    )
+    random_vertex["markers"]["C"] = 10
+    scores = polarization_scores_component_graph(
+        graph, component_id="PXLCMP0000000", n_permutations=10, random_seed=1
+    )
     # We don't expect to get a value for B, since it has only one value in it.
     # Hence it is filtered out.
     expected = pd.DataFrame.from_dict(