+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Changed name of TCRb to TCRVb5 antibody in human-immunology-panel file and bumped to version 0.5.0.
 * Renaming of component metrics in adata
 * Use MPX graph compatible permutation strategy when calculating Moran's I.
+* Marker filtering is now done after count transformation in polarization score calculation.
 
 ### Fixed